+++ src/drvizer/__init__.py
@@ -5,7 +5,9 @@ Genomic Visualization Tools
 __version__ = "1.0.0"
 
 from .api import DrViz
+from ._parallel import ParallelCoverageError
 
 __all__ = [
     'DrViz',
+    'ParallelCoverageError',
 ]